# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/quote": {
      "post": {
        "description": "Creates a quote in `draft` status. The body must carry at least one line; lines are immutable for the rest of the quote lifecycle.",
        "operationId": "createQuote",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateQuoteInputDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Quote stored in draft and ready to be sent.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuoteOutputDto"
                }
              }
            }
          },
          "400": {
            "description": "Body validation failed (see error payload)."
          }
        },
        "summary": "Create draft quote with lines",
        "tags": [
          "Quote · Commands"
        ]
      }
    },
    "/quote/{id}/send": {
      "post": {
        "description": "Transitions the quote from `draft` to `sent` and queues an email notification for the given recipient.\n\nThe email itself is dispatched asynchronously via the outbox + EventBridge → SQS pipeline; this endpoint returns as soon as the state transition is committed and the event is enqueued. Delivery is eventual.\n\nOnce sent, the quote is frozen and the customer can `accept` or `reject` it.",
        "operationId": "sendQuote",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Quote UUID v7.",
            "schema": {
              "example": "019ad951-368a-7de5-b7ba-add19cfd187b",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendQuoteInputDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Quote frozen and notification dispatch queued.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuoteOutputDto"
                }
              }
            }
          },
          "400": {
            "description": "Body validation failed (see error payload)."
          },
          "404": {
            "description": "Quote not found."
          },
          "409": {
            "description": "Quote is not in draft status."
          }
        },
        "summary": "Send draft quote to a recipient",
        "tags": [
          "Quote · Commands"
        ]
      }
    },
    "/quote/{id}/accept": {
      "post": {
        "description": "Marks a `sent` quote as `accepted` on behalf of the customer. Terminal state — the quote can no longer transition.",
        "operationId": "acceptQuote",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Quote UUID v7.",
            "schema": {
              "example": "019ad951-368a-7de5-b7ba-add19cfd187b",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Quote accepted (terminal state).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuoteOutputDto"
                }
              }
            }
          },
          "404": {
            "description": "Quote not found."
          },
          "409": {
            "description": "Quote is not in sent status."
          }
        },
        "summary": "Accept sent quote",
        "tags": [
          "Quote · Commands"
        ]
      }
    },
    "/quote/{id}/reject": {
      "post": {
        "description": "Marks a `sent` quote as `rejected` on behalf of the customer. Terminal state — the quote can no longer transition.",
        "operationId": "rejectQuote",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Quote UUID v7.",
            "schema": {
              "example": "019ad951-368a-7de5-b7ba-add19cfd187b",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Quote rejected (terminal state).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuoteOutputDto"
                }
              }
            }
          },
          "404": {
            "description": "Quote not found."
          },
          "409": {
            "description": "Quote is not in sent status."
          }
        },
        "summary": "Reject sent quote",
        "tags": [
          "Quote · Commands"
        ]
      }
    },
    "/quote/{id}": {
      "get": {
        "description": "Returns the full quote: header (id, customer, currency, status), every line, and the computed total in cents.",
        "operationId": "getQuote",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Quote UUID v7.",
            "schema": {
              "example": "019ad951-368a-7de5-b7ba-add19cfd187b",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Full quote returned.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuoteOutputDto"
                }
              }
            }
          },
          "404": {
            "description": "Quote not found."
          }
        },
        "summary": "Get full quote by id",
        "tags": [
          "Quote · Queries"
        ]
      }
    },
    "/quote/{id}/summary": {
      "get": {
        "description": "Compact projection for list views: `id`, `customerId`, `status`, `currency`, and the total in cents. Lines are **not** included — the total is aggregated in SQL so this endpoint avoids loading per-line rows.",
        "operationId": "getQuoteSummary",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Quote UUID v7.",
            "schema": {
              "example": "019ad951-368a-7de5-b7ba-add19cfd187b",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Quote summary returned.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuoteSummaryOutputDto"
                }
              }
            }
          },
          "404": {
            "description": "Quote not found."
          }
        },
        "summary": "Get quote summary by id",
        "tags": [
          "Quote · Queries"
        ]
      }
    },
    "/healthz": {
      "get": {
        "description": "Liveness and readiness probe in a single endpoint.\n\n### Guarantees\n- The NestJS process is running and the HTTP stack responds.\n- Every registered background worker has ticked within its own threshold.\n\n### Status codes\n- **200 OK** — every worker is healthy.\n- **503 Service Unavailable** — at least one worker has not ticked within its threshold.\n\nIntended for Kubernetes `livenessProbe` and `readinessProbe`.",
        "operationId": "HealthzController_health",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Every worker is healthy.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthOutputDto"
                }
              }
            }
          },
          "503": {
            "description": "At least one worker is unhealthy."
          }
        },
        "summary": "Health check",
        "tags": [
          "Healthz"
        ]
      }
    }
  },
  "info": {
    "title": "@org/api",
    "description": "Boilerplate for NestJS projects",
    "version": "1.0.0",
    "contact": {}
  },
  "tags": [],
  "servers": [],
  "components": {
    "schemas": {
      "QuoteLineInputDto": {
        "type": "object",
        "properties": {
          "description": {
            "type": "string",
            "description": "Line description",
            "example": "Consulting — 8 hours"
          },
          "quantity": {
            "type": "number",
            "description": "Quantity (integer ≥ 1)",
            "example": 2
          },
          "unitPriceCents": {
            "type": "number",
            "description": "Unit price in the Quote currency, expressed in cents",
            "example": 15000
          }
        },
        "required": [
          "description",
          "quantity",
          "unitPriceCents"
        ]
      },
      "CreateQuoteInputDto": {
        "type": "object",
        "properties": {
          "customerId": {
            "type": "string",
            "description": "Customer identifier (UUID)",
            "example": "019be5e6-c357-733b-8857-25fef291279f"
          },
          "currency": {
            "type": "string",
            "description": "ISO 4217 currency code (3 letters, uppercase)",
            "example": "EUR"
          },
          "lines": {
            "description": "At least one line. Lines are immutable once the quote is created.",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/QuoteLineInputDto"
            }
          }
        },
        "required": [
          "customerId",
          "currency",
          "lines"
        ]
      },
      "QuoteLineOutputDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "019be5e6-c357-733b-8857-25fef291279f"
          },
          "description": {
            "type": "string",
            "example": "Consulting — 8 hours"
          },
          "quantity": {
            "type": "number",
            "example": 2
          },
          "unitPriceCents": {
            "type": "number",
            "description": "Unit price in cents",
            "example": 15000
          },
          "lineTotalCents": {
            "type": "number",
            "description": "Computed: unitPriceCents × quantity",
            "example": 30000
          }
        },
        "required": [
          "id",
          "description",
          "quantity",
          "unitPriceCents",
          "lineTotalCents"
        ]
      },
      "QuoteOutputDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "019ad951-368a-7de5-b7ba-add19cfd187b"
          },
          "customerId": {
            "type": "string",
            "example": "019be5e6-c357-733b-8857-25fef291279f"
          },
          "currency": {
            "type": "string",
            "example": "EUR"
          },
          "status": {
            "type": "string",
            "enum": [
              "draft",
              "sent",
              "accepted",
              "rejected"
            ],
            "example": "draft"
          },
          "lines": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/QuoteLineOutputDto"
            }
          },
          "totalAmountCents": {
            "type": "number",
            "description": "Sum of all line totals (cents)",
            "example": 30000
          }
        },
        "required": [
          "id",
          "customerId",
          "currency",
          "status",
          "lines",
          "totalAmountCents"
        ]
      },
      "SendQuoteInputDto": {
        "type": "object",
        "properties": {
          "recipient": {
            "type": "string",
            "description": "Email address that will receive the quote notification.",
            "example": "customer@example.com"
          }
        },
        "required": [
          "recipient"
        ]
      },
      "QuoteSummaryOutputDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "019ad951-368a-7de5-b7ba-add19cfd187b"
          },
          "customerId": {
            "type": "string",
            "example": "019be5e6-c357-733b-8857-25fef291279f"
          },
          "currency": {
            "type": "string",
            "example": "EUR"
          },
          "status": {
            "type": "string",
            "enum": [
              "draft",
              "sent",
              "accepted",
              "rejected"
            ],
            "example": "sent"
          },
          "totalAmountCents": {
            "type": "number",
            "description": "Aggregated server-side via SQL SUM",
            "example": 35000
          }
        },
        "required": [
          "id",
          "customerId",
          "currency",
          "status",
          "totalAmountCents"
        ]
      },
      "QuoteSentEventDto": {
        "type": "object",
        "properties": {}
      },
      "WorkerHealthSnapshotDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Worker name (e.g. outbox-relay).",
            "example": "outbox-relay"
          },
          "healthy": {
            "type": "boolean",
            "description": "True when this worker ticked within its threshold.",
            "example": true
          },
          "lastTickAt": {
            "type": "string",
            "description": "Timestamp of the last tick, or null before the first one.",
            "example": "2026-04-24T10:30:00.000Z",
            "nullable": true
          },
          "thresholdMs": {
            "type": "number",
            "description": "Max age (ms) the last tick can have before this worker is considered unhealthy.",
            "example": 10000
          }
        },
        "required": [
          "name",
          "healthy",
          "thresholdMs"
        ]
      },
      "HealthOutputDto": {
        "type": "object",
        "properties": {
          "healthy": {
            "type": "boolean",
            "description": "True when every registered worker is healthy.",
            "example": true
          },
          "workers": {
            "description": "Snapshot of every registered worker.",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WorkerHealthSnapshotDto"
            }
          }
        },
        "required": [
          "healthy",
          "workers"
        ]
      }
    }
  }
}